fix stack overflow on recursive export trie
When the export trie is malformed and is recursive, this Exported::parse()
will stack overflow when trying to parse it.

Detect this condition and return an error instead. This is implemented by
remembering the list of parent nodes that we visited in the trie, and erroring
if we ever try to visit the same node a second time.

I initially tried to implement it by ensuring that the Cursor always makes
forward progress when parsing the trie, but some valid binaries, e.g. /usr/bin/smbd
layout their trie in a way that requires to sometime jump backwards to get the
child nodes.